% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirene.R
\name{get_nearby}
\alias{get_nearby}
\title{Get all firms nearby a point}
\usage{
get_nearby(latitude, longitude,
  endpoint = "https://entreprise.data.gouv.fr/api/sirene/v1")
}
\arguments{
\item{latitude}{a double}

\item{longitude}{a double}

\item{endpoint}{an url}
}
\value{
a tibble
}
\description{
Get all firms nearby a point
}
\examples{
get_nearby(latitude = 48.8503, longitude = 2.30831)

}
