% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirene.R
\name{get_unitelegale}
\alias{get_unitelegale}
\title{Get unite legale}
\usage{
get_unitelegale(
  siren,
  endpoint = "https://entreprise.data.gouv.fr/api/sirene/"
)
}
\arguments{
\item{siren}{a valid siren number (9 digits)}

\item{endpoint}{}
}
\value{
a tibble
}
\description{
Get unite legale
}
\examples{
\dontrun{
get_unitelegale(siren = "382357721")
}

}
