% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensirene.R
\name{get_url}
\alias{get_url}
\title{Get url of Sirene API User interface}
\usage{
get_url(siret, endpoint = "https://entreprise.data.gouv.fr")
}
\arguments{
\item{siret}{a valid siret number as a string}

\item{endpoint}{API endpoint}
}
\value{
a string
}
\description{
Get url of Sirene API User interface
}
\examples{

get_url(siret = "38235772100028")

\dontrun{
browseURL(get_url(siret = "38235772100028"))
}

}
